fix: move realtime types out of typescript declaration files. There are not shipped when build.
